+++ gauge.go
@@ -5,33 +5,36 @@ type gaugeImpl struct {
 	collector *collector
 }
 
-func (g *gaugeImpl) Increment() {
-	g.IncrementBy(1)
+func (g *gaugeImpl) Increment(labels ...MetricLabel) {
+	g.IncrementBy(1, labels...)
 }
 
-func (g *gaugeImpl) IncrementBy(by float64) {
+func (g *gaugeImpl) IncrementBy(by float64, labels ...MetricLabel) {
 	g.collector.mutex.Lock()
 	defer g.collector.mutex.Unlock()
 
-	value := g.collector.get(g.name, map[string]string{})
-	g.collector.set(g.name, map[string]string{}, value+by)
+	realLabels := metricLabels(labels).toMap()
+	value := g.collector.get(g.name, realLabels)
+	g.collector.set(g.name, realLabels, value+by)
 }
 
-func (g *gaugeImpl) Decrement() {
-	g.DecrementBy(1)
+func (g *gaugeImpl) Decrement(labels ...MetricLabel) {
+	g.DecrementBy(1, labels...)
 }
 
-func (g *gaugeImpl) DecrementBy(by float64) {
+func (g *gaugeImpl) DecrementBy(by float64, labels ...MetricLabel) {
 	g.collector.mutex.Lock()
 	defer g.collector.mutex.Unlock()
 
-	value := g.collector.get(g.name, map[string]string{})
-	g.collector.set(g.name, map[string]string{}, value-by)
+	realLabels := metricLabels(labels).toMap()
+	value := g.collector.get(g.name, realLabels)
+	g.collector.set(g.name, realLabels, value-by)
 }
 
-func (g *gaugeImpl) Set(value float64) {
+func (g *gaugeImpl) Set(value float64, labels ...MetricLabel) {
 	g.collector.mutex.Lock()
 	defer g.collector.mutex.Unlock()
 
-	g.collector.set(g.name, map[string]string{}, value)
+	realLabels := metricLabels(labels).toMap()
+	g.collector.set(g.name, realLabels, value)
 }